# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Edu_Connect API
  version: 1.0.0
  description: OpenAPI specification for the Edu_Connect backend (TSUAP scaffold)
servers:
  - url: http://localhost:4000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
          format: float
        provider:
          type: string
        reference:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /api/payments/initiate:
    post:
      summary: Initiate payment (authenticated)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                provider:
                  type: string
                reference:
                  type: string
              required: [amount, provider, reference]
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /api/payments/webhook:
    post:
      summary: Payment provider webhook (idempotent, HMAC signed)
      parameters:
        - in: header
          name: x-webhook-signature
          schema:
            type: string
          required: true
          description: HMAC SHA256 signature of the raw request body using PAYMENT_WEBHOOK_SECRET
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Bad request or signature verification failed
        '409':
          description: Duplicate webhook (already processed)
...